prototype/requirements.txt
typer[all]
openai>=1.0.0
python-dotenv
InquirerPy
pytest
psycopg2-binary
pinecone-client